+++ CHANGELOG.md
@@ -9,8 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [7.2.0] - 2022-08-02
 
 ### Added
-- Add "role" command and "exists" subcommand
+- Add "role" command and "exists", "memberships" subcommands
   [cyberark/cyberark-conjur-cli#417](https://github.com/cyberark/cyberark-conjur-cli/pull/417)
+  [cyberark/cyberark-conjur-cli#419](https://github.com/cyberark/cyberark-conjur-cli/pull/419)
 - Add "resource" command and "exists" subcommand
   [cyberark/cyberark-conjur-cli#418](https://github.com/cyberark/cyberark-conjur-cli/pull/418)
 - Add "show" command